# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/media/{mediaId}/access": {
      "get": {
        "operationId": "MediaController_createAccess",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {},
                    "meta": {
                      "$ref": "#/components/schemas/ApiResponseMetaDto"
                    }
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ],
        "tags": [
          "Media"
        ]
      }
    },
    "/api/v1/media/{mediaId}/content": {
      "get": {
        "operationId": "MediaController_getContent",
        "parameters": [
          {
            "$ref": "#/components/parameters/RequestId"
          },
          {
            "name": "mediaId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "expires",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 2147483647,
              "type": "number"
            }
          },
          {
            "name": "signature",
            "required": true,
            "in": "query",
            "schema": {
              "pattern": "^[a-f0-9]{64}$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Raw media bytes; errors still use the JSON error envelope.",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/ErrorEnvelope"
          }
        },
        "tags": [
          "Media"
        ]
      }
    },
...
  },
  "info": {
    "title": "HSK System API",
    "description": "Generated from the backend code. Envelope, errors and request id: docs/api/api.md §1.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ApiResponseMetaDto": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "requestId": {
            "type": "string",
            "description": "Same value as the `X-Request-ID` response header."
          }
        },
        "required": [
          "timestamp",
          "requestId"
        ]
      },
...
      "ApiErrorDto": {
        "type": "object",
        "properties": {
          "details": {
            "type": "object",
            "additionalProperties": true
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable code, for example `UNAUTHORIZED`."
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiErrorEnvelopeDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResponseMetaDto"
          }
        },
        "required": [
          "success",
          "error",
          "meta"
        ]
      }
    },
    "parameters": {
      "RequestId": {
        "name": "X-Request-ID",
        "in": "header",
        "required": false,
        "description": "UUID or 32 hex characters; any other value is replaced by a new UUID. Echoed in the X-Request-ID response header and meta.requestId.",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "ErrorEnvelope": {
        "description": "Error envelope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorEnvelopeDto"
            }
          }
        }
      }
    }
  }
}